lambda/dev/hello_world/requirements.txt
boto3
python-telegram-bot
